refactor(web): split page.tsx into five modules to satisfy monolith policy
Split apps/web/app/page.tsx (1467 → 460 lines) into five sibling modules,
each ≤500 lines, to clear the last monolith-check failure:

- panel-types.ts        — type/interface defs + TOKEN_KEY
- command-format.ts     — command/result formatting + commandExamples
- job-helpers.ts        — job/doctor data helpers + savedMessage
- panel-components.tsx  — presentational React components + status utils
- use-panel-data.ts     — useState/useMemo/useEffect + async handlers hook

Pure structural extraction, no behavior change. Deduped a duplicate
formatBytes into command-format.ts and tightened the ArtifactHandle type
to carry kind/display_path/line_count.

Verified: npx tsc --noEmit (0 errors), enforce_monoliths.py (all 6 pass),
cargo check (0 errors). Version bump 5.8.0 → 5.8.1 (chore=patch).
