+++ apps/web/openapi/axon.json
@@ -10,7 +10,7 @@
       "name": "MIT",
       "identifier": "MIT"
     },
-    "version": "5.8.0"
+    "version": "5.8.1"
   },
   "paths": {
     "/v1/ask": {
