+++ apps/web/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@axon/admin-panel",
-  "version": "5.8.0",
+  "version": "5.8.1",
   "private": true,
   "scripts": {
     "build": "next build",
